refactor(browser): extract design-mode picker script to its own module
Move the ~270-line design-mode picker driver (the in-page IIFE injected
via executeJavaScript) out of the inline template literal in browserIpc.js
into web/electron/src/designModeScript.js, so it lints and highlights as
its own file instead of an opaque backtick string.

Behavior is byte-identical: the function is moved verbatim, keeping its
(nonce) signature and internal SELECT/SUBMIT/DISMISS marker derivation, so
the produced script string matches the old one exactly for the same nonce
(verified by diffing the output across several nonces). browserIpc.js now
imports buildDesignModeScript and re-exports it, so the existing tests that
require it from browserIpc keep working unchanged. No security logic
touched — the per-enable nonce, gesture gate, and console-marker channel
are all preserved as-is.
